==> code/3-07_prec_emp_descriptive_report_weighted.Rmd <==
---
title: "Precarious employment - descriptive analysis (weighted)"
author: "Andrew Pulford"
date: ""
output: word_document
always_allow_html: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

## remove any existing objects from global environment
rm(list=ls()) 

## stop scientific notation
options(scipen=999, digits = 1)

### load packages
library(tidyverse)
library(DiagrammeR)
library(knitr)
library(TraMineR)
#remotes::install_github("rstudio/webshot2")
library(poLCA)
library(magick)
library(tinytex)
#webshot::install_phantomjs()

##### load dataframes ----------------------------------------------------------

### master raw data --------
# This includes selected variables for all individual respondents across Understanding Society waves 3-10
master_raw <- readRDS("../raw_data/master_raw1.rds")

## waves 3-6
master_raw_a <- master_raw %>% filter(wv_n %in% c(3:6))

## waves 7-10
master_raw_b <- master_raw %>% filter(wv_n %in% c(7:10))


## kepp only indivs with valid end point -- relates to node A in flowchart --
end_spine_a <- master_raw_a %>% 
  filter(wv_n==6) %>% 
  dplyr::select(pidp)

end_spine_b <- master_raw_b %>% 
  filter(wv_n==10) %>% 
  dplyr::select(pidp)

#end_spine_b_alt <- master_raw_b_alt %>% 
#  filter(wv_n==10) %>% 
#  dplyr::select(pidp)

end_valid_a <- end_spine_a %>% 
  left_join(master_raw_a)

end_valid_b <- end_spine_b %>% 
  left_join(master_raw_b)

#end_valid_b_alt <- end_spine_b_alt %>% 
#  left_join(master_raw1b_alt)

master_indivs_a <- length(unique(end_valid_a$pidp))
master_indivs_b <- length(unique(end_valid_b$pidp))
#master_indivs_b_alt <- length(unique(master_raw_b_alt$pidp))


### no valid weight ---------
non_weight_spine_a <- readRDS("../look_ups/no_weight_spine_a.rds")
non_weight_spine_b <- readRDS("../look_ups/no_weight_spine_b.rds")

### non-working age ---------
non_working_age_spine_a <- readRDS("../look_ups/non_working_age_spine_a.rds")
non_working_age_spine_b <- readRDS("../look_ups/non_working_age_spine_b.rds")

### no valid case at end-point ----------
#no_endpoint_a <- readRDS("../look_ups/no_endpoint_a.rds")
#no_endpoint_b <- readRDS("../look_ups/no_endpoint_b.rds")

### eligible population ---------
# This includes individual respondents aged 16-64; Understanding Society waves 3-10 
eligible_a <- readRDS("../working_data/eligible_pop_a.rds") 
eligible_b <- readRDS("../working_data/eligible_pop_b.rds") 

### censored cases ---------
censor_combined_a <- readRDS("../look_ups/censor_combined_a.rds")
censor_combined_b <- readRDS("../look_ups/censor_combined_b.rds")

### analytic samples  ---------
dfas1a <- read_rds("../analytic_sample_data/dfas1a.rds")
dfas1b <- read_rds("../analytic_sample_data/dfas1b.rds")

##### flowchart nodes A to xx --------------------------------------------------

### Node A:  ----------------
### Total number of individuals in end wave of study 
## number of unique individuals (node A)
raw_unique_indiv_a <- master_indivs_a
raw_unique_indiv_b <- master_indivs_b

### Node AA: ----------------
### number of individuals excluded due to having no valid weight, not being 
### working-age or no valid response at study endpoint

## no valid weight
no_weight_a <- nrow(non_weight_spine_a)
no_weight_b <- nrow(non_weight_spine_b)

## non-working age
non_work_age_a <- nrow(non_working_age_spine_a)
non_work_age_b <- nrow(non_working_age_spine_b)
#non_work_age <- prettyNum(non_work_age, big.mark=",")

## no valid endpoint response
#no_valid_end_a <- length(unique(no_endpoint_a$pidp))
#no_valid_end_b <- length(unique(no_endpoint_b$pidp))
#no_wv2 <- prettyNum(no_wv2, big.mark=",")


### Node B: ----------------
### number of individuals in eligible population
eligible_unique_indiv_a <- length(unique(eligible_a$pidp))
eligible_unique_indiv_b <- length(unique(eligible_b$pidp))
#eligible_descr_unique_indiv <- prettyNum(eligible_descr_unique_indiv, big.mark=",")


### Node BB: ----------------
### number of individuals censored due to retirement or incomplete

## retired
retired_spine_a <- censor_combined_a %>% 
  filter(censor_reason=="retired")

retired_spine_b <- censor_combined_b %>% 
  filter(censor_reason=="retired")

censor_retired_a <- nrow(retired_spine_a)
censor_retired_b <- nrow(retired_spine_b)
#censor_retired <- prettyNum(censor_retired, big.mark=",")


## number of individuals censored due to incomplete data
incomplete_spine_a <- censor_combined_a %>% 
  filter(censor_reason=="incomplete")

incomplete_spine_b <- censor_combined_b %>% 
  filter(censor_reason=="incomplete")

incomplete_a <- nrow(incomplete_spine_a)
incomplete_b <- nrow(incomplete_spine_b)
#censor_lost <- prettyNum(censor_lost, big.mark=",")


### Node C: ---------------- 
### Number of individuals and observations in analytic samples

dfas1a_observations <- nrow(dfas1a)
dfas1b_observations <- nrow(dfas1b)
#dfas1_observations <- prettyNum(dfas1_observations, big.mark=",")

dfas1a_unique_indivs <- length(unique(dfas1a$pidp))
dfas1b_unique_indivs <- length(unique(dfas1b$pidp))
#dfas1_unique_indivs <- prettyNum(dfas1_unique_indivs, big.mark=",")


##### Study endpoint sample characteristics
sample_chars_endpoint <- readRDS("../working_data/weighted/sample_chars_endpoint.rds")  %>% 
  mutate(measure=tolower(measure))

## formatting for table
# for estimates set 1dp
sample_chars_endpoint$est <- formatC(sample_chars_endpoint$est, digits = 1, format = "f")

# for number of cases set 1,000 separator
sample_chars_endpoint$n <- prettyNum(sample_chars_endpoint$n, big.mark=",")

# for number of cases set <10 to -999 so can suppress low numbers
sample_chars_endpoint <- sample_chars_endpoint %>% 
  mutate(n = ifelse(n%in%c(1:10),-999,n))




sample_chars_endpoint_wide <- sample_chars_endpoint %>% 
  dplyr::select(-se) %>% 
  filter(wv_n==6) %>% 
  rename("n6" = "n",
         "est6" = "est") %>% 
  ungroup() %>% 
  dplyr::select(-wv_n)

temp <- sample_chars_endpoint %>% 
  filter(wv_n==10) %>% 
  rename("n10" = "n",
         "est10" = "est") %>% 
  ungroup() %>% 
  dplyr::select(-c(wv_n,se))

sample_chars_endpoint_wide <- sample_chars_endpoint_wide %>% 
  full_join(temp, by = c("var", "measure"))

# for number of cases convert to character so NAs can be removed
sample_chars_endpoint_wide$n6 <- as.character(sample_chars_endpoint_wide$n6)
sample_chars_endpoint_wide$n10 <- as.character(sample_chars_endpoint_wide$n10)

sample_chars_endpoint_wide <- sample_chars_endpoint_wide %>% 
  group_by(var) %>% 
  mutate(row_n = row_number()) %>% 
  mutate(var = ifelse(row_n==1,var,"")) %>% 
  ungroup() %>% 
  mutate(n6 = ifelse(n6=="NA","",
                    ifelse(n6=="-999","-",n6)),
         n10 = ifelse(n10=="NA","",
                    ifelse(n10=="-999","-",n10))) %>% 
  dplyr::select(-row_n)

sample_chars_endpoint_wide$measure <- str_to_sentence(sample_chars_endpoint_wide$measure)

### Node D: -------------
### Exposure groupings

### load class spines and keep only exposure classes
## employment contract LCA spine
emp_contracta_5class_spine <- readRDS("../working_data/emp_contracta_5class_spine.rds") %>% 
  filter(emp_contract_class == "non-permanent employment")


emp_contractb_5class_spine <- readRDS("../working_data/emp_contractb_5class_spine.rds") %>% 
  filter(emp_contract_class == "non-permanent employment")

### employment spells LCA spine
emp_spellsa_3class_spine <- readRDS("../working_data/emp_spellsa_3class_spine.rds") %>% 
  filter(emp_spells_class == "broken employment")

emp_spellsb_3class_spine <- readRDS("../working_data/emp_spellsb_3class_spine.rds") %>% 
  filter(emp_spells_class == "broken employment")

### multiple employment LCA spine
multi_empa_5class_spine <- readRDS("../working_data/multi_empa_5class_spine.rds") %>% 
  filter(multi_emp_class == "multiple employment")

multi_empb_5class_spine <- readRDS("../working_data/multi_empb_5class_spine.rds") %>% 
  filter(multi_emp_class == "multiple employment")


# non-permanent class sample A 
non_perm_a <- nrow(emp_contracta_5class_spine)

# non_perm_b
non_perm_b <- nrow(emp_contractb_5class_spine)

# broken_a
broken_a <-  nrow(emp_spellsa_3class_spine)

# broken_b
broken_b <-  nrow(emp_spellsb_3class_spine)

# multi_a
multi_a <- nrow(multi_empa_5class_spine)

# multi_b
multi_b <- nrow(multi_empb_5class_spine)

#### Node E: -------------
### Outcome caseness

## self-rated health

# events
srh_case <- "x"
srh_case_a <- dfas1a %>% filter(wv_n==6 & srh_bin=="fair/poor")
srh_case_a <- srh_case_a %>% nrow()
  
srh_case_b <- dfas1b %>% filter(wv_n==10 & srh_bin=="fair/poor")
srh_case_b <- srh_case_b %>% nrow()

# non-events
srh_nocase <- "x"
srh_nocase_a <- dfas1a %>% filter(wv_n==6 & srh_bin=="excellent/very good/good")
srh_nocase_a <- srh_nocase_a %>% nrow()
  
srh_nocase_b <- dfas1b %>% filter(wv_n==10 & srh_bin=="excellent/very good/good")
srh_nocase_b <- srh_nocase_b %>% nrow()
## GHQ-12
# events
ghq_case <- "x"

ghq_case_a <- dfas1a %>% filter(wv_n==6 & ghq_case4=="4 or more")
ghq_case_a <- ghq_case_a %>% nrow()

ghq_case_b <- dfas1b %>% filter(wv_n==10 & ghq_case4=="4 or more")
ghq_case_b <- ghq_case_b %>% nrow()

# non-events
ghq_nocase <- "x"
ghq_nocase_a <- dfas1a %>% filter(wv_n==6 & ghq_case4=="0-3")
ghq_nocase_a <- ghq_nocase_a %>% nrow()

ghq_nocase_b <- dfas1b %>% filter(wv_n==10 & ghq_case4=="0-3")
ghq_nocase_b <- ghq_nocase_b %>% nrow()

#### weights spines -------------------
weight_spine_a <- readRDS("../look_ups/weights_spine_a.rds") %>% 
  dplyr::select(pidp, weight_flag, indinub_xw) %>% 
  filter(weight_flag==1) %>% 
  dplyr::select(-weight_flag)

weight_spine_b <- readRDS("../look_ups/weights_spine_b.rds") %>% 
  dplyr::select(pidp, weight_flag, indinui_xw) %>% 
  filter(weight_flag==1) %>% 
  dplyr::select(-weight_flag)

#### age-sex standardised prevalence dataframes
## emp contract
svy_dsr_df1 <- readRDS("../working_data/weighted/svy_dsr_df1.rds")

## emp spells
svy_dsr_df2 <- readRDS("../working_data/weighted/svy_dsr_df2.rds")

## multiple emp
svy_dsr_df3 <- readRDS("../working_data/weighted/svy_dsr_df3.rds")




```
# Abstract

## Background

## Methods

## Results

## Conclusions

\newpage
# Introduction
xxxxxx

\newpage
# Methods
## Data source
xxxxxxxx

## Measurements
xxxxxxxx

## Statistical analysis
xxxxxxxx

\newpage
# Results
## Sample description
A total of `r prettyNum(raw_unique_indiv_a, big.mark=",")` individuals participated in UKHLS wave 6 and `r prettyNum(raw_unique_indiv_b, big.mark=",")` in wave 10 (Figure 1). We excluded `r prettyNum(non_work_age_a, big.mark=",")` and `r prettyNum(non_work_age_b, big.mark=",")` individuals respectively for not being aged 20-64 at the end point of the study period; plus `r prettyNum(no_weight_a, big.mark=",")` and `r prettyNum(no_weight_b, big.mark=",")` individuals respectively who did not have valid weights. From eligible populations of `r prettyNum(eligible_unique_indiv_a, big.mark=",")` and `r prettyNum(eligible_unique_indiv_b, big.mark=",")` we excluded `r prettyNum(censor_retired_a, big.mark=",")` and `r prettyNum(censor_retired_b, big.mark=",")` retired individuals; and `r prettyNum(incomplete_a, big.mark=",")` and  `r prettyNum(incomplete_b, big.mark=",")` individuals who reported missing data for key exposure and outcomes of interest at waves 6 or 10 respectively. Table 1 presents key characteristics of the analytic sample at baseline by employment contract status (UKHLS wave 2). Responses to all four waves in the study period were provided by 96.5% of analytic sample A and 95.7% of analytic sample B (Supplementary figure 1). Valid response sequences ascending in order of frequency are presented in Supplementary figure 2.

\newpage
**Figure 1: Flowcharts for inclusion and exclusion of UKHLS respondents to create analytic sample (unweighted)**

**(a) Sample A (waves 3-6)**
```{r, echo=FALSE, include=TRUE, fig.width=8}


## flowchart A
grViz(paste0("digraph {
  graph []
  node []
    A [label = '",prettyNum(raw_unique_indiv_a, big.mark=",")," individuals' , shape = box, 
    width = 3, height = 1]
    AA [label = '",prettyNum(non_work_age_a, big.mark=",")," non-working age
    
    ",prettyNum(no_weight_a, big.mark=",")," no valid weight' , shape = box, 
    width = 3, height = 1]
    B [label = '",prettyNum(eligible_unique_indiv_a, big.mark=",")," eligible individuals' , shape = box, 
    width = 3, height = 1]
    BB [label = '",prettyNum(censor_retired_a, big.mark=",")," retired
    
    ",prettyNum(incomplete_a, big.mark=",")," incomplete data', shape = box, width = 3, height = 1]
    C [label = '",prettyNum(dfas1a_unique_indivs, big.mark=",")," in analytic sample A', shape = box, 
    width = 3, height = 1]
#    D [label = '",prettyNum(non_perm_a, big.mark=",")," in non-permanent employment class
#    
#    ",prettyNum(broken_a, big.mark=",")," in broken employment class
#    
#    ",prettyNum(multi_a, big.mark=",")," in multiple employment class' , shape = box, 
#    width = 3, height = 1]
#    E [label = '",prettyNum(srh_case_a, big.mark=",")," cases - poor self-reported health
#    
#    ",prettyNum(srh_nocase_a, big.mark=",")," event free - poor self-reported health
#    
#    ",prettyNum(ghq_case_a, big.mark=",")," cases - common mental health disorder
#    
#    ",prettyNum(ghq_nocase_a, big.mark=",")," event free - common mental health disorder' , shape = box, #width = 3, height = 1]
  edge []
    A->B
    A->AA
    B->BB
    B->C
#    C->D
#    D->E
{ rank = same; A; AA }
{ rank = same; B; BB }}"))
```

**(b) Sample B (waves 7-10)**
```{r, echo=FALSE, include=TRUE, fig.width=8}


## flowchart B
grViz(paste0("digraph {
  graph []
  node []
    A [label = '",prettyNum(raw_unique_indiv_b, big.mark=",")," individuals', shape = box, width = 3, height = 1]
    AA [label = '",prettyNum(non_work_age_b, big.mark=",")," non-working age
    
    ",prettyNum(no_weight_b, big.mark=",")," no valid weight' , shape = box, 
    width = 3, height = 1]
    B [label = '",prettyNum(eligible_unique_indiv_b, big.mark=",")," eligible individuals' , shape = box, 
    width = 3, height = 1]
    BB [label = '",prettyNum(censor_retired_b, big.mark=",")," retired
    
    ",prettyNum(incomplete_b, big.mark=",")," incomplete data', shape = box, width = 3, height = 1]
    C [label = '",prettyNum(dfas1b_unique_indivs, big.mark=",")," in analytic sample B', shape = box, 
    width = 3, height = 1]
#    D [label = '",prettyNum(non_perm_b, big.mark=",")," in non-permanent employment class
#    
#    ",prettyNum(broken_b, big.mark=",")," in broken employment class
#    
#    ",prettyNum(multi_b, big.mark=",")," in multiple employment class' , shape = box, 
#    width = 3, height = 1]
#    E [label = '",prettyNum(srh_case_b, big.mark=",")," cases - poor self-reported health
#    
#    ",prettyNum(srh_nocase_b, big.mark=",")," event free - poor self-reported health
#    
#    ",prettyNum(ghq_case_b, big.mark=",")," cases - common mental health disorder
#    
#    ",prettyNum(ghq_nocase_b, big.mark=",")," event free - common mental health disorder' , shape = box, #width = 3, height = 1]
  edge []
    A->B
    A->AA
    B->BB
    B->C
#    C->D
#    D->E
{ rank = same; A; AA }
{ rank = same; B; BB }}"))
```

```{r, echo=FALSE}
## add checks to ensure number of individuals adds up
check_AB_a <- eligible_unique_indiv_a+non_work_age_a+no_weight_a==raw_unique_indiv_a
check_BC_a <- censor_retired_a+incomplete_a+dfas1a_unique_indivs==eligible_unique_indiv_a

check_AB_b <- eligible_unique_indiv_b+non_work_age_b+no_weight_b==raw_unique_indiv_b
check_BC_b <- censor_retired_b+incomplete_b+dfas1b_unique_indivs==eligible_unique_indiv_b

if(check_AB_a!=TRUE){print("Sample A - boxes AA and B to not equal box A")}
if(check_BC_a!=TRUE){print("Sample A - boxes BB and C to not equal box B")}

if(check_AB_b!=TRUE){print("Sample B - boxes AA and B to not equal box A")}
if(check_BC_b!=TRUE){print("Sample B - boxes BB and C to not equal box B")}

```
\newpage
**Table 1: Analytic sample characteristics at study endpoint (UKHLS waves 6 and 10)**
```{r, echo=FALSE}

#sample_chars_endpoint_wide <- sample_chars_endpoint_wide[,c(1:10)]
kable(sample_chars_endpoint_wide, 
      col.names = c("Variable", 
                    "Measure", 
                    "n (wave 6)", 
                    "% (wave 6)", 
                    "n (wave 10)", 
                    "% (wave 10)"), 
      format.args = list(big.mark = ","))

```
## Sequence analysis of precarious employment exposures of interest
```{r, echo=FALSE, include=FALSE}

#### ---------------------------------------------------------------------------
#### Employment contract
#### ---------------------------------------------------------------------------

### recode employment status variables to create an employment contract variable
dfas1a_seq <- dfas1a

dfas1b_seq <- dfas1b 

### wide format for creating sequence data
dfas1a_seq_wide  <-  dfas1a_seq %>% 
  dplyr::select(pidp,wv_n,emp_contract) %>% 
  mutate(wv=paste0("wv_",wv_n)) %>% 
  dplyr::select(-wv_n) %>% 
  pivot_wider(names_from = wv, values_from = emp_contract, values_fill = "missing") %>% 
  filter(wv_6!="missing") #%>% 
#  dplyr::select(-wv_NA)


dfas1b_seq_wide  <-  dfas1b_seq %>% 
  dplyr::select(pidp,wv_n,emp_contract) %>% 
  mutate(wv=paste0("wv_",wv_n)) %>% 
  dplyr::select(-wv_n) %>% 
  pivot_wider(names_from = wv, values_from = emp_contract, values_fill = "missing") %>% 
  filter(wv_10!="missing")

### add on weights
dfas1a_seq_wide <- dfas1a_seq_wide %>% 
  left_join(weight_spine_a)

dfas1b_seq_wide <- dfas1b_seq_wide %>% 
  left_join(weight_spine_b)

### define labels and codes for sequence analysis
## retaining missing values for now but plan to imput
emp_contract_labs <- c("fixed term", "missing", "permanent", "unemployed/not in employment" )
emp_contract_code <- c("FT", "NA", "PE", "UE")

### create sequence data
emp_contract.seq.a <- seqdef(dfas1a_seq_wide, 2:5, states = emp_contract_code,
                             labels = emp_contract_labs, weights = dfas1a_seq_wide$indinub_xw)

emp_contract.seq.b <- seqdef(dfas1b_seq_wide, 2:5, states = emp_contract_code,
                             labels = emp_contract_labs, weights = dfas1b_seq_wide$indinui_xw)


```

Supplementary figure 3 presents employment contract sequences ascending in order of frequency. Consistent permanent employment was the most commonly observed sequence across both samples, followed by consistent un/non-employment. A large minority of cases had employment contract sequences with varying states during the study periods. Supplementary figure 4 presents employment spell sequences ascending in order of frequency. Consistent unbroken employment was the most commonly observed sequence across both samples, followed by consistent un/non-employment. A large minority of cases had employment spell sequences with varying states during the study periods. Supplementary figure 5 presents multiple employment sequences ascending in order of frequency. Consistent single employment was the most commonly observed sequence across both samples, followed by consistent un/non-employment. A large minority of cases had sequences with varying states during the study periods.  

## Latent class analysis

```{r, echo=FALSE}
class_pc <- readRDS("../working_data/weighted/class_pc.rds")

### weighted % for exposure classes ----------
## employment contract
non_perm_a_pc <- class_pc$pc[class_pc$measure=="non-permanent employment" &
                             class_pc$sample_grp=="A"]
non_perm_b_pc <- class_pc$pc[class_pc$measure=="non-permanent employment" &
                             class_pc$sample_grp=="B"]

## employment spells
broken_a_pc <- class_pc$pc[class_pc$measure=="broken employment" &
                             class_pc$sample_grp=="A"]
broken_b_pc <- class_pc$pc[class_pc$measure=="broken employment" &
                             class_pc$sample_grp=="B"]



## multiple employment
multi_a_pc <- class_pc$pc[class_pc$measure=="multiple employment" &
                             class_pc$sample_grp=="A"]
multi_b_pc <- class_pc$pc[class_pc$measure=="multiple employment" &
                             class_pc$sample_grp=="B"]


## emp contract fit stats
empcontracta_fit_stats <- read.csv("../output/descriptive/empcontracta_lca_fit_stats.csv")


```
### Employment contract
Seven latent class models were run specifying between two and eight latent classes. Based on the Bayesian Information Criterion (BIC) scores for these models, a five class solution appeared to be the optimum model (Supplementary table 1). Visual inspection of the BIC elbow plot indicated that suitable "joints" in the BIC score occurred at three and five solutions (Supplementary figure 6). Visual inspection of class membership indicated that the five class model had greater face validity than the three class model (Supplementary figure 7). Supplementary figure 8 presents employment contract sequences ascending in order of frequency grouped by latent class membership for Sample A. The five latent classes have been codified as: (a) individuals who were in permanent employment for most or all the study period; (b) individuals who were unemployed/no in employment for most or all of the study period; (c) individuals who moved into employment during the study period; (d) individuals who moved out of employment during the study period; and (e) individuals who were often in non-permanent employment during the study period. Individuals who were in non-permanent employment for most or all the study period represented `r prettyNum(non_perm_a_pc)`% of Sample A.

Latent class analysis of employment contract across study period B provided a very similar solution to that modelled for study period A, with a five class solution presenting as the optimum solution (Supplementary table 2, Supplementary figure 9, Supplementary figure 10). Supplementary figure 11 presents employment contract sequences ascending in order of frequency grouped by latent class membership for Sample B. The same codification of latent classes identified for Sample A was applicable to Sample B. Individuals who were in non-permanent employment for most or all the study period represented `r prettyNum(non_perm_b_pc)`% of Sample B.

### Broken employment spells
Latent class analysis of broken employment spells between survey waves identified a three class model as providing the optimum solution (Supplementary table 3, Supplementary figure 12, Supplementary figure 13). Supplementary figure 14 presents broken employment sequences ascending in order of frequency grouped by latent class membership for Sample A. The three latent classes have been codified as: (a) individuals who were in unbroken employment for most or all the study period; (b) individuals who were unemployed/no in employment for most or all of the study period; (c) individuals who reported broken employment spells for most or all the study waves. Individuals who reported broken employment spells for most or all the study waves represented `r prettyNum(broken_a_pc)`% of Sample A.

Latent class analysis of broken employment spells across study period B provided a very similar solution to that modelled for study period A, with a five class solution presenting as the optimum solution (Supplementary table 4, Supplementary figure 15, Supplementary figure 16). Supplementary figure 17 presents employment spells sequences ascending in order of frequency grouped by latent class membership for Sample B. Individuals who reported broken employment spells for most or all the study waves represented `r prettyNum(broken_b_pc)`% of Sample B.

### Multiple employment
```{r, echo=FALSE}

multi_empa_lca_fit_stats <- read.csv("../output/descriptive/multi_empa_lca_fit_stats.csv")
```
Latent class analysis of multiple employment across study period A provided a very similar solution to that modelled for employment contract, with a five class model presenting as the optimum solution (Supplementary table 5, Supplementary figure 18, Supplementary figure 19). 

Supplementary figure 20 presents multiple employment sequences ascending in order of frequency grouped by latent class membership for Sample A. The five latent classes have been codified as: (a) individuals who were in single employment for most or all the study period; (b) individuals who were unemployed/no in employment for most or all of the study period; (c) individuals who moved into employment during the study period; (d) individuals who moved out of employment during the study period; and (e) individuals who were often in multiple employment during the study period. Individuals who reported multiple employment for most or all the study waves represented `r prettyNum(multi_a_pc)`% of Sample A.

```{r, echo=FALSE}

multi_empb_lca_fit_stats <- read.csv("../output/descriptive/multi_empb_lca_fit_stats.csv")
```
Latent class analysis of multiple employment across study period B provided a very similar solution to that modelled study period A, with a five class solution presenting as the optimum solution (Supplementary table 6, Supplementary figure 21, Supplementary figure 22). Supplementary figure 23 presents multiple employment sequences ascending in order of frequency grouped by latent class membership for Sample A. Individuals who reported multiple employment for most or all the study waves represented `r prettyNum(multi_b_pc)`% of Sample B.

## Outcomes of interest by exposure class
### Self-rated health
```{r, echo=FALSE}
#### load in logistic regression df for self-rated health
srh_facet_df <- readRDS("../working_data/weighted/srh_facet_df.rds")

#### extract data for in-line code

### employment contract
## sample A
# OR
srh_non_perm_or_a <- as.character(srh_facet_df$est[srh_facet_df$measure=="non-permanent employment" &
                     srh_facet_df$sample_grp=="A"])
srh_non_perm_or_a <- str_sub(srh_non_perm_or_a, start = 1, end = 4)
# LCI
srh_non_perm_lci_a <- as.character(srh_facet_df$lci[srh_facet_df$measure=="non-permanent employment" &
                     srh_facet_df$sample_grp=="A"])
srh_non_perm_lci_a <- str_sub(srh_non_perm_lci_a, start = 1, end = 4)
# UCI
srh_non_perm_uci_a <- as.character(srh_facet_df$uci[srh_facet_df$measure=="non-permanent employment" &
                     srh_facet_df$sample_grp=="A"])
srh_non_perm_uci_a <- str_sub(srh_non_perm_uci_a, start = 1, end = 4)

## sample B
# OR
srh_non_perm_or_b <- as.character(srh_facet_df$est[srh_facet_df$measure=="non-permanent employment" &
                     srh_facet_df$sample_grp=="B"])
srh_non_perm_or_b <- str_sub(srh_non_perm_or_b, start = 1, end = 4)
# LCI
srh_non_perm_lci_b <- as.character(srh_facet_df$lci[srh_facet_df$measure=="non-permanent employment" &
                     srh_facet_df$sample_grp=="B"])
srh_non_perm_lci_b <- str_sub(srh_non_perm_lci_b, start = 1, end = 4)
# UCI
srh_non_perm_uci_b <- as.character(srh_facet_df$uci[srh_facet_df$measure=="non-permanent employment" &
                     srh_facet_df$sample_grp=="B"])
srh_non_perm_uci_b <- str_sub(srh_non_perm_uci_b, start = 1, end = 4)

### employment spells
## sample A
# OR
srh_broken_or_a <- as.character(srh_facet_df$est[srh_facet_df$measure=="broken employment" &
                     srh_facet_df$sample_grp=="A"])
srh_broken_or_a <- str_sub(srh_broken_or_a, start = 1, end = 4)
# LCI
srh_broken_lci_a <- as.character(srh_facet_df$lci[srh_facet_df$measure=="broken employment" &
                     srh_facet_df$sample_grp=="A"])
srh_broken_lci_a <- str_sub(srh_broken_lci_a, start = 1, end = 4)
# UCI
srh_broken_uci_a <- as.character(srh_facet_df$uci[srh_facet_df$measure=="broken employment" &
                     srh_facet_df$sample_grp=="A"])
srh_broken_uci_a <- str_sub(srh_broken_uci_a, start = 1, end = 4)

## sample B
# OR
srh_broken_or_b <- as.character(srh_facet_df$est[srh_facet_df$measure=="broken employment" &
                     srh_facet_df$sample_grp=="B"])
srh_broken_or_b <- str_sub(srh_broken_or_b, start = 1, end = 4)
# LCI
srh_broken_lci_b <- as.character(srh_facet_df$lci[srh_facet_df$measure=="broken employment" &
                     srh_facet_df$sample_grp=="B"])
srh_broken_lci_b <- str_sub(srh_broken_lci_b, start = 1, end = 4)
# UCI
srh_broken_uci_b <- as.character(srh_facet_df$uci[srh_facet_df$measure=="broken employment" &
                     srh_facet_df$sample_grp=="B"])
srh_broken_uci_b <- str_sub(srh_broken_uci_b, start = 1, end = 4)

### multiple employment
## sample A
# OR
srh_multi_or_a <- as.character(srh_facet_df$est[srh_facet_df$measure=="multiple employment" &
                     srh_facet_df$sample_grp=="A"])
srh_multi_or_a <- str_sub(srh_multi_or_a, start = 1, end = 4)
# LCI
srh_multi_lci_a <- as.character(srh_facet_df$lci[srh_facet_df$measure=="multiple employment" &
                     srh_facet_df$sample_grp=="A"])
srh_multi_lci_a <- str_sub(srh_multi_lci_a, start = 1, end = 4)
# UCI
srh_multi_uci_a <- as.character(srh_facet_df$uci[srh_facet_df$measure=="multiple employment" &
                     srh_facet_df$sample_grp=="A"])
srh_multi_uci_a <- str_sub(srh_multi_uci_a, start = 1, end = 4)

## sample B
# OR
srh_multi_or_b <- as.character(srh_facet_df$est[srh_facet_df$measure=="multiple employment" &
                     srh_facet_df$sample_grp=="B"])
srh_multi_or_b <- str_sub(srh_multi_or_b, start = 1, end = 4)
# LCI
srh_multi_lci_b <- as.character(srh_facet_df$lci[srh_facet_df$measure=="multiple employment" &
                     srh_facet_df$sample_grp=="B"])
srh_multi_lci_b <- str_sub(srh_multi_lci_b, start = 1, end = 4)
# UCI
srh_multi_uci_b <- as.character(srh_facet_df$uci[srh_facet_df$measure=="multiple employment" &
                     srh_facet_df$sample_grp=="B"])
srh_multi_uci_b <- str_sub(srh_multi_uci_b, start = 1, end = 4)
```

The odds of poor self-rated health were not found to differ for members of the persistent non-permanent class in either study period (`r srh_non_perm_or_a` OR, `r srh_non_perm_lci_a` to `r srh_non_perm_uci_a` 95% CI in Sample A; `r srh_non_perm_or_b` OR, `r srh_non_perm_lci_b` to `r srh_non_perm_uci_b` 95% CI in Sample B) compared to the persistently permanently employed class (Figure 2). The odds for poor self-rated health were found to be higher among members of the broken employment spells class in both study periods (`r srh_broken_or_a` OR, `r srh_broken_lci_a` to `r srh_broken_uci_a` 95% CI in Sample A; `r srh_broken_or_b` OR, `r srh_broken_lci_b` to `r srh_broken_uci_b` 95% CI in Sample B) compared to the unbroken employment spells class. The odds of poor self-rated health were lower for the persistent multiple employment class in both study periods compared to the persistent single employment class(`r srh_multi_or_a` OR, `r srh_multi_lci_a` to `r srh_multi_uci_a` 95% CI in Sample A; `r srh_multi_or_b` OR, `r srh_multi_lci_b` to `r srh_multi_uci_b` 95% CI in Sample B).

\newpage
**Figure 2: The odds of poor self-rate health by precarious employment history class compared to more securely employed reference group (study periods A and B) **

```{r, echo=FALSE}
image_read("../output/weighted/srh_prev_grouped_facet.tiff")

```


### Common mental disorders

```{r, echo=FALSE}
#### load in logistic regression df for self-rated health
ghq_facet_df <- readRDS("../working_data/weighted/ghq_facet_df.rds")

#### extract data for in-line code

### employment contract
## sample A
# OR
ghq_non_perm_or_a <- as.character(ghq_facet_df$est[ghq_facet_df$measure=="non-permanent employment" &
                     ghq_facet_df$sample_grp=="A"])
ghq_non_perm_or_a <- str_sub(ghq_non_perm_or_a, start = 1, end = 4)
# LCI
ghq_non_perm_lci_a <- as.character(ghq_facet_df$lci[ghq_facet_df$measure=="non-permanent employment" &
                     ghq_facet_df$sample_grp=="A"])
ghq_non_perm_lci_a <- str_sub(ghq_non_perm_lci_a, start = 1, end = 4)
# UCI
ghq_non_perm_uci_a <- as.character(ghq_facet_df$uci[ghq_facet_df$measure=="non-permanent employment" &
                     ghq_facet_df$sample_grp=="A"])
ghq_non_perm_uci_a <- str_sub(ghq_non_perm_uci_a, start = 1, end = 4)

## sample B
# OR
ghq_non_perm_or_b <- as.character(ghq_facet_df$est[ghq_facet_df$measure=="non-permanent employment" &
                     ghq_facet_df$sample_grp=="B"])
ghq_non_perm_or_b <- str_sub(ghq_non_perm_or_b, start = 1, end = 4)
# LCI
ghq_non_perm_lci_b <- as.character(ghq_facet_df$lci[ghq_facet_df$measure=="non-permanent employment" &
                     ghq_facet_df$sample_grp=="B"])
ghq_non_perm_lci_b <- str_sub(ghq_non_perm_lci_b, start = 1, end = 4)
# UCI
ghq_non_perm_uci_b <- as.character(ghq_facet_df$uci[ghq_facet_df$measure=="non-permanent employment" &
                     ghq_facet_df$sample_grp=="B"])
ghq_non_perm_uci_b <- str_sub(ghq_non_perm_uci_b, start = 1, end = 4)

### employment spells
## sample A
# OR
ghq_broken_or_a <- as.character(ghq_facet_df$est[ghq_facet_df$measure=="broken employment" &
                     ghq_facet_df$sample_grp=="A"])
ghq_broken_or_a <- str_sub(ghq_broken_or_a, start = 1, end = 4)
# LCI
ghq_broken_lci_a <- as.character(ghq_facet_df$lci[ghq_facet_df$measure=="broken employment" &
                     ghq_facet_df$sample_grp=="A"])
ghq_broken_lci_a <- str_sub(ghq_broken_lci_a, start = 1, end = 4)
# UCI
ghq_broken_uci_a <- as.character(ghq_facet_df$uci[ghq_facet_df$measure=="broken employment" &
                     ghq_facet_df$sample_grp=="A"])
ghq_broken_uci_a <- str_sub(ghq_broken_uci_a, start = 1, end = 4)

## sample B
# OR
ghq_broken_or_b <- as.character(ghq_facet_df$est[ghq_facet_df$measure=="broken employment" &
                     ghq_facet_df$sample_grp=="B"])
ghq_broken_or_b <- str_sub(ghq_broken_or_b, start = 1, end = 4)
# LCI
ghq_broken_lci_b <- as.character(ghq_facet_df$lci[ghq_facet_df$measure=="broken employment" &
                     ghq_facet_df$sample_grp=="B"])
ghq_broken_lci_b <- str_sub(ghq_broken_lci_b, start = 1, end = 4)
# UCI
ghq_broken_uci_b <- as.character(ghq_facet_df$uci[ghq_facet_df$measure=="broken employment" &
                     ghq_facet_df$sample_grp=="B"])
ghq_broken_uci_b <- str_sub(ghq_broken_uci_b, start = 1, end = 4)

### multiple employment
## sample A
# OR
ghq_multi_or_a <- as.character(ghq_facet_df$est[ghq_facet_df$measure=="multiple employment" &
                     ghq_facet_df$sample_grp=="A"])
ghq_multi_or_a <- str_sub(ghq_multi_or_a, start = 1, end = 4)
# LCI
ghq_multi_lci_a <- as.character(ghq_facet_df$lci[ghq_facet_df$measure=="multiple employment" &
                     ghq_facet_df$sample_grp=="A"])
ghq_multi_lci_a <- str_sub(ghq_multi_lci_a, start = 1, end = 4)
# UCI
ghq_multi_uci_a <- as.character(ghq_facet_df$uci[ghq_facet_df$measure=="multiple employment" &
                     ghq_facet_df$sample_grp=="A"])
ghq_multi_uci_a <- str_sub(ghq_multi_uci_a, start = 1, end = 4)

## sample B
# OR
ghq_multi_or_b <- as.character(ghq_facet_df$est[ghq_facet_df$measure=="multiple employment" &
                     ghq_facet_df$sample_grp=="B"])
ghq_multi_or_b <- str_sub(ghq_multi_or_b, start = 1, end = 4)
# LCI
ghq_multi_lci_b <- as.character(ghq_facet_df$lci[ghq_facet_df$measure=="multiple employment" &
                     ghq_facet_df$sample_grp=="B"])
ghq_multi_lci_b <- str_sub(ghq_multi_lci_b, start = 1, end = 4)
# UCI
ghq_multi_uci_b <- as.character(ghq_facet_df$uci[ghq_facet_df$measure=="multiple employment" &
                     ghq_facet_df$sample_grp=="B"])
ghq_multi_uci_b <- str_sub(ghq_multi_uci_b, start = 1, end = 4)
```

Membership of the persistent non-permanent employment class was associated with higher odds of GHQ-12 caseness compared to the persistent permanent employment class in study period A (`r ghq_non_perm_or_a` OR, `r ghq_non_perm_lci_a` to `r ghq_non_perm_uci_a` 95% CI) but not in study period B (`r ghq_non_perm_or_b` OR, `r ghq_non_perm_lci_b` to `r ghq_non_perm_uci_b` 95% CI) (Figure 3). The odds of GHQ-12 caseness were higher among members of the broken employment spells class compared to the unbroken spells class in study period A (`r ghq_broken_or_a` OR, `r ghq_broken_lci_a` to `r ghq_broken_uci_a` 95% CI) but not in study period B (`r ghq_broken_or_b` OR, `r ghq_broken_lci_b` to `r ghq_broken_uci_b` 95% CI). Membership of the persistent multiple employment class was associated with similar odds of GHQ-12 caseness in study period A (`r ghq_multi_or_a` OR, `r ghq_multi_lci_a` to `r ghq_multi_uci_a` 95% CI) and higher odds in study period B (`r ghq_multi_or_b` OR, `r ghq_multi_lci_b` to `r ghq_multi_uci_b` 95 CI) compared to the persistent single employment class.

\newpage
**Figure 3: The odds of common mental disorder by precarious employment history class compared to more securely employed reference group (study periods A and B) **

```{r, echo=FALSE}
image_read("../output/weighted/ghq_prev_grouped_facet.tiff")

```


# Discussion
xxxxxx

## Implications
xxxxxx

## Strengths and limitations
xxxxxxx

# Conclusions
xxxxx

\newpage
# Supplementary materials
## Missing data (unweighted)
```{r, include=FALSE, echo=FALSE}
### total number of cases in each wave 16-64
## waves 3-6
spine_total_a <- dfas1a %>% group_by(wv_n) %>% 
  summarise(n=n()) %>% 
  ungroup() 

# number and % of individuals in sample based on endpoint
indivs_a <- spine_total_a[[4,2]]

spine_total_a <- spine_total_a %>% 
  mutate(total = indivs_a) %>% 
  mutate(pc = n/total*100) %>% 
  mutate(as = "a")

## waves7-10
spine_total_b <- dfas1b %>% group_by(wv_n) %>% 
  summarise(n=n()) %>% 
  ungroup() 

# number and % of individuals in sample based on endpoint
indivs_b <- spine_total_b[[4,2]]

spine_total_b <- spine_total_b %>% 
  mutate(total = indivs_b) %>% 
  mutate(pc = n/total*100) %>% 
  mutate(as = "b")

spine_total <- spine_total_a %>% bind_rows(spine_total_b)

### sequence analysis - wave non-response

## specify labels and codes (same for both samples)
spine_labs <- c("response", "missing")
spine_code <- c("RESP", "NA")


## wide format for sequencing data
# waves 3-6
spine_wide_a <- dfas1a %>%
  dplyr::select(pidp,wv_n) %>% 
  mutate(response=1) %>% 
  mutate(wv=paste0("wv_",wv_n)) %>% 
  dplyr::select(-wv_n) %>% 
  pivot_wider(names_from = wv, values_from = response, values_fill = 99)

# waves 7-10
spine_wide_b <- dfas1b %>%
  dplyr::select(pidp,wv_n) %>% 
  mutate(response=1) %>% 
  mutate(wv=paste0("wv_",wv_n)) %>% 
  dplyr::select(-wv_n) %>% 
  pivot_wider(names_from = wv, values_from = response, values_fill = 99)


spine.seq.a <- seqdef(spine_wide_a, 2:5, states = spine_code,
                    labels = spine_labs)
spine.seq.b <- seqdef(spine_wide_b, 2:5, states = spine_code,
                    labels = spine_labs)


### total number of cases by number of waves reported
## waves 3-6
spine_waves_a <- dfas1a %>% 
  group_by(pidp) %>% 
  summarise(n_wvs = n()) %>% # calculate number of waves responded to
  ungroup() %>% 
  group_by(n_wvs) %>% 
  summarise(n = n()) %>% 
  ungroup() %>% 
  mutate(pc= n/sum(n)*100) %>% 
  mutate(as = "Sample A (waves 3-6)")


## waves 7-10
spine_waves_b <- dfas1b %>% 
  group_by(pidp) %>% 
  summarise(n_wvs = n()) %>% # calculate number of waves responded to
  ungroup() %>% 
  group_by(n_wvs) %>% 
  summarise(n = n()) %>% 
  ungroup() %>% 
  mutate(pc= n/sum(n)*100) %>% 
  mutate(as = "Sample B (waves 7-10")

```

**Supplementary figure 1: Number of waves responded to by percentage of the analytic sample  (unweighted)**
```{r, echo=FALSE, fig.width=4,fig.height=3}
## join together
spine_waves <- spine_waves_a %>% bind_rows(spine_waves_b)

# chart
spine_waves %>% 
  ggplot(aes(x=n_wvs, y=pc)) +
  geom_col() +
#  scale_x_continuous(limits = c(0,5),breaks = 1:4) +
  xlab("Number of waves responded to") +
  ylab("Percentage of cases") +
  theme_classic() +
  facet_wrap(~as, scales = "free")


```

\newpage
## Sequence analysis of precarious employment exposures of interest
**Supplementary figure 2: Valid response sequences ascending in order of frequency  (unweighted)**

```{r, echo=FALSE, fig.width=12,fig.height=15}

# sequence frequency plot (all common sequences)
#png("output/descriptive/seqfplot_missing.png", width = 960, height = 1920)
par(mfrow=c(1,3))
seqfplot(spine.seq.a, 
         idxs=1:900, # to add more lines
         with.legend = F, 
         border = NA, 
         main = "Sample A (waves 3-6)")
seqfplot(spine.seq.b, 
         idxs=1:900, # to add more lines
         with.legend = F, 
         border = NA, 
         main = "Sample B (waves 7-10)")
seqlegend(spine.seq.a, cex = 1.3)
#dev.off()
```

\newpage


**Supplementary figure 3: Employment contract sequences ascending in order of frequency**
```{r, echo=FALSE, fig.width=9,fig.height=9}

#
# sequence frequency plot (all common sequences)
par(mfrow=c(1,3))
seqfplot(emp_contract.seq.a, 
         idxs=1:900, # to add more lines
         with.legend = F, 
         border = NA, 
         main = "Sample A (waves 3-6)")
seqfplot(emp_contract.seq.b, 
         idxs=1:900, # to add more lines
         with.legend = F, 
         border = NA, 
         main = "Sample B (waves 7-10)")
seqlegend(emp_contract.seq.a, cex = 1.3)


```

```{r, echo=FALSE, include=FALSE}
#### ---------------------------------------------------------------------------
#### Broken employment
#### ---------------------------------------------------------------------------

### employment spells -----
# create numeric version
dfas1a_seq2 <- dfas1a

dfas1b_seq2 <- dfas1b 

### wide format for creating sequence data
dfas1a_seq_wide2  <-  dfas1a_seq2 %>% 
  dplyr::select(pidp,wv_n,broken_emp) %>% 
  mutate(wv=paste0("wv_",wv_n)) %>% 
  #  mutate(broken_emp = ifelse(is.na(broken_emp),"missing",broken_emp)) %>% 
  dplyr::select(-wv_n) %>% 
  pivot_wider(names_from = wv, values_from = broken_emp, values_fill = "missing") %>% 
  filter(wv_6!="missing") #%>% 
#  dplyr::select(-wv_NA)


dfas1b_seq_wide2  <-  dfas1b_seq2 %>% 
  dplyr::select(pidp,wv_n,broken_emp) %>% 
  mutate(broken_emp = ifelse(is.na(broken_emp),"missing",broken_emp)) %>% 
  mutate(wv=paste0("wv_",wv_n)) %>% 
  dplyr::select(-wv_n) %>% 
  pivot_wider(names_from = wv, values_from = broken_emp, values_fill = "missing")


### add on weights
dfas1a_seq_wide2 <- dfas1a_seq_wide2 %>% 
  left_join(weight_spine_a)

dfas1b_seq_wide2 <- dfas1b_seq_wide2 %>% 
  left_join(weight_spine_b)

### define labels and codes for sequence analysis
## retaining missing values for now but plan to impute
broken_emp_labs <- c("Broken employment", "Missing", "No employment spells", "Unbroken employment" )
broken_emp_code <- c("BE","NA", "NE", "UE")

### create sequence data
broken_emp.seq.a <- seqdef(dfas1a_seq_wide2, 2:5, states = broken_emp_code,
                           labels = broken_emp_labs, weights = dfas1a_seq_wide2$indinub_xw)

broken_emp.seq.b <- seqdef(dfas1b_seq_wide2, 2:5, states = broken_emp_code,
                           labels = broken_emp_labs, weights = dfas1b_seq_wide$indinui_xw)


```

**Supplementary figure 4: Employment spell sequences ascending in order of frequency**
```{r, echo=FALSE, fig.width=9,fig.height=9}
# sequence frequency plot (all common sequences)
#png("output/descriptive/seqfplot_emp_contract.png", width = 960, height = 1920)
par(mfrow=c(1,3))
seqfplot(broken_emp.seq.a, 
         idxs=1:900, # to add more lines
         with.legend = F, 
         border = NA, 
         main = "Sample A (waves 3-6)")
seqfplot(broken_emp.seq.b, 
         idxs=1:900, # to add more lines
         with.legend = F, 
         border = NA, 
         main = "Sample A (waves 7-10)")
seqlegend(broken_emp.seq.a, cex = 1.3)

```

```{r, echo=FALSE, include=FALSE}

#### ---------------------------------------------------------------------------
#### Multiple jobs
#### ---------------------------------------------------------------------------

dfas1a_seq3 <- dfas1a
dfas1b_seq3 <- dfas1b

### wide format for creating sequence data
dfas1a_seq_wide3  <-  dfas1a_seq3 %>% 
  dplyr::select(pidp,wv_n,j2has_dv) %>% 
  mutate(wv=paste0("wv_",wv_n)) %>% 
  dplyr::select(-wv_n) %>% 
  pivot_wider(names_from = wv, values_from = j2has_dv, values_fill = "missing")

dfas1b_seq_wide3  <-  dfas1b_seq3 %>% 
  dplyr::select(pidp,wv_n,j2has_dv) %>% 
  mutate(wv=paste0("wv_",wv_n)) %>% 
  dplyr::select(-wv_n) %>% 
  pivot_wider(names_from = wv, values_from = j2has_dv, values_fill = "missing")

### define labels and codes for sequence analysis
## retaining missing values for now but plan to impute
multi_jobs_labs <- c("Missing", "No", "Yes")
multi_jobs_code <- c("NA","N", "Y")

### create sequence data
multi_jobs.seq.a <- seqdef(dfas1a_seq_wide3, 2:5, states = multi_jobs_code,
                           labels = multi_jobs_labs)

multi_jobs.seq.b <- seqdef(dfas1b_seq_wide3, 2:5, states = multi_jobs_code,
                           labels = multi_jobs_labs)

```


**Supplementary figure 5: Multiple job sequences ascending in order of frequency**
```{r, echo=FALSE, fig.width=9,fig.height=9}
# sequence frequency plot (all common sequences)
#png("output/descriptive/seqfplot_emp_contract.png", width = 960, height = 1920)
par(mfrow=c(1,3))
seqfplot(multi_jobs.seq.a, 
         idxs=1:900, # to add more lines
         with.legend = F, 
         border = NA, 
         main = "Sample A (waves 3-6)")
seqfplot(multi_jobs.seq.b, 
         idxs=1:900, # to add more lines
         with.legend = F, 
         border = NA, 
         main = "Sample B (waves 7-10)")
seqlegend(multi_jobs.seq.a, cex = 1.3)
```

\newpage

## Latent class analysis - Employment contract
### Employment contract - Sample A
**Supplementary table 1: Model fit statistics for employment contract latent class analysis solutions (2-8 classes) - Sample A **

```{r, echo=FALSE}
kable(empcontracta_fit_stats)
```

\newpage

**Supplementary figure 6: BIC elbow plot for employment contract latent class analysis solutions (2-8 classes) - Sample A **

```{r, echo=FALSE}
image_read("../output/descriptive/empcontracta_lca_elbow.tiff")


```

```{r, echo=FALSE}
#readRDS("../output/descriptive/emp_contracta_lca_final_model.rds")
```

\newpage

**Supplementary figure 7: Employment contract latent class membership for five class model - Sample A **

```{r, echo=FALSE}
image_read("../output/descriptive/emp_contracta_lca_final.tiff")
```

\newpage

**Supplementary figure 8: Employment contract sequences ascending in order of frequency grouped by latent class membership - Sample A **

```{r, echo=FALSE}
image_read("../output/weighted/emp_contracta_grouped_seqfplot.tiff")

```

\newpage

### Employment contract - Sample B
```{r, echo=FALSE}

empcontractb_fit_stats <- read.csv("../output/descriptive/empcontractb_lca_fit_stats.csv")
```

**Supplementary table 2: Model fit statistics for employment contract latent class analysis solutions (2-8 classes) - Sample B** 
```{r, echo=FALSE}
kable(empcontractb_fit_stats)
```

**Supplementary figure 9: BIC elbow plot for employment contract latent class analysis solutions (2-8 classes) - Sample B**
```{r, echo=FALSE}
image_read("../output/descriptive/empcontractb_lca_elbow.tiff")
```

\newpage

**Supplementary figure 10: Employment contract sequences ascending in order of frequency grouped by latent class membership - Sample B **
```{r, echo=FALSE}
image_read("../output/weighted/emp_contractb_grouped_seqfplot.tiff")
```

**Supplementary figure 11: Employment contract latent class membership for five class model - Sample B  **

```{r, echo=FALSE}
image_read("../output/descriptive/emp_contractb_lca_final.tiff")
```

\newpage

## Latent class analysis - Broken employment spells
### Broken employment spells - Sample A
**Supplementary table 3: Model fit statistics for employment spells latent class analysis solutions (2-8 classes) - Sample A** 
```{r, echo=FALSE}

## emp spells fit stats A
empspellsa_fit_stats <- read.csv("../output/descriptive/empspellsa_lca_fit_stats.csv")

kable(empspellsa_fit_stats)
```

\newpage



**Supplementary figure 12: BIC elbow plot for employment spells latent class analysis solutions (2-8 classes) - Sample A** 
```{r, echo=FALSE}
image_read("../output/descriptive/empspellsa_lca_elbow.tiff")


```

**Supplementary figure 13: Employment spells latent class membership for three class model - Sample A**

```{r, echo=FALSE}
image_read("../output/descriptive/emp_spellsa_lca_final.tiff")
```

**Supplementary figure 14: Employment spells sequences ascending in order of frequency grouped by latent class membership - Sample A **

```{r, echo=FALSE}
image_read("../output/weighted/emp_spellsa_grouped_seqfplot.tiff")

```


\newpage

### Broken employment spells - Sample B
**Supplementary table 4: Model fit statistics for employment spells latent class analysis solutions (2-8 classes) - Sample B** 
```{r, echo=FALSE}

## emp spells fit stats B
empspellsb_fit_stats <- read.csv("../output/descriptive/empspellsb_lca_fit_stats.csv")

kable(empspellsb_fit_stats)
```

**Supplementary figure 15: BIC elbow plot for employment spells latent class analysis solutions (2-8 classes) - Sample B** 
```{r, echo=FALSE}
image_read("../output/descriptive/empspellsb_lca_elbow.tiff")


```

**Supplementary figure 16: Employment spells latent class membership for three class model - Sample B**
```{r, echo=FALSE}
image_read("../output/descriptive/emp_spellsa_lca_final.tiff")

```

**Supplementary figure 17: Employment spells sequences ascending in order of frequency grouped by latent class membership - Sample B **

```{r, echo=FALSE}
image_read("../output/weighted/emp_spellsb_grouped_seqfplot.tiff")

```

\newpage

## Latent class analysis - Multiple employment
### Multiple employment - Sample A
**Supplementary table 5: Model fit statistics for multiple employment latent class analysis solutions (2-8 classes) - Sample A **

```{r, echo=FALSE}
kable(multi_empa_lca_fit_stats)
```

\newpage

**Supplementary figure 18: BIC elbow plot for multiple employment latent class analysis solutions (2-8 classes) - Sample A **

```{r, echo=FALSE}
image_read("../output/descriptive/multi_empa_lca_elbow.tiff")


```

\newpage

```{r, echo=FALSE}
#readRDS("../output/descriptive/multi_empa_lca_final_model.rds")
```

**Supplementary figure 19: Multiple employment latent class membership for five class model - Sample A **

```{r, echo=FALSE}
image_read("../output/descriptive/multi_empa_lca_final.tiff")
```

\newpage

**Supplementary figure 20: Multiple employment sequences ascending in order of frequency grouped by latent class membership - Sample A **

```{r, echo=FALSE}
image_read("../output/weighted/emp_contracta_grouped_seqfplot.tiff")

```

\newpage


### Multiple employment - Sample B
**Supplementary table 6: Model fit statistics for multiple employment latent class analysis solutions (2-8 classes) - Sample B **

```{r, echo=FALSE}
kable(multi_empb_lca_fit_stats)
```

\newpage

**Supplementary figure 21: BIC elbow plot for multiple employment latent class analysis solutions (2-8 classes) - Sample B **

```{r, echo=FALSE}
image_read("../output/descriptive/multi_empb_lca_elbow.tiff")


```


```{r, echo=FALSE}
#readRDS("../output/descriptive/multi_empb_lca_final_model.rds")
```

\newpage

**Supplementary figure 22: Multiple employment latent class membership for five class model - Sample B **

```{r, echo=FALSE}
image_read("../output/descriptive/multi_empb_lca_final.tiff")
```

**Supplementary figure 23: Multiple employment sequences ascending in order of frequency grouped by latent class membership - Sample B **

```{r, echo=FALSE}
image_read("../output/weighted/emp_contractb_grouped_seqfplot.tiff")

```

\newpage

## Prevalence of outcomes of interest by employment contract class
### Employment contract

```{r, echo=FALSE, include=FALSE}
### % poor SRH
## non-permanent sample A
non_perm_srh_prev_a <- svy_dsr_df1$value[svy_dsr_df1$sample_grp=="A" &
                                       svy_dsr_df1$outcome_lab=="poor self-rated health" &
                                       svy_dsr_df1$sex_dv=="both" &
                                       svy_dsr_df1$class_mem=="Non-permanent employment"]

## non-permanent sample B
non_perm_srh_prev_b <- svy_dsr_df1$value[svy_dsr_df1$sample_grp=="B" &
                                       svy_dsr_df1$outcome_lab=="poor self-rated health" &
                                       svy_dsr_df1$sex_dv=="both" &
                                       svy_dsr_df1$class_mem=="Non-permanent employment"]

## permanent sample A
perm_srh_prev_a <- svy_dsr_df1$value[svy_dsr_df1$sample_grp=="A" &
                                       svy_dsr_df1$outcome_lab=="poor self-rated health" &
                                       svy_dsr_df1$sex_dv=="both" &
                                       svy_dsr_df1$class_mem=="Permanent employment"]

## permanent sample B
perm_srh_prev_b <- svy_dsr_df1$value[svy_dsr_df1$sample_grp=="B" &
                                       svy_dsr_df1$outcome_lab=="poor self-rated health" &
                                       svy_dsr_df1$sex_dv=="both" &
                                       svy_dsr_df1$class_mem=="Permanent employment"]


### % poor mental health
## non-permanent sample A
non_perm_ghq_prev_a <- svy_dsr_df1$value[svy_dsr_df1$sample_grp=="A" &
                                       svy_dsr_df1$outcome_lab=="poor mental health" &
                                       svy_dsr_df1$sex_dv=="both" &
                                       svy_dsr_df1$class_mem=="Non-permanent employment"]

## non-permanent sample B
non_perm_ghq_prev_b <- svy_dsr_df1$value[svy_dsr_df1$sample_grp=="B" &
                                       svy_dsr_df1$outcome_lab=="poor mental health" &
                                       svy_dsr_df1$sex_dv=="both" &
                                       svy_dsr_df1$class_mem=="Non-permanent employment"]

## permanent sample A
perm_ghq_prev_a <- svy_dsr_df1$value[svy_dsr_df1$sample_grp=="A" &
                                       svy_dsr_df1$outcome_lab=="poor mental health" &
                                       svy_dsr_df1$sex_dv=="both" &
                                       svy_dsr_df1$class_mem=="Permanent employment"]

## permanent sample B
perm_ghq_prev_b <- svy_dsr_df1$value[svy_dsr_df1$sample_grp=="B" &
                                       svy_dsr_df1$outcome_lab=="poor mental health" &
                                       svy_dsr_df1$sex_dv=="both" &
                                       svy_dsr_df1$class_mem=="Permanent employment"]


```

The non-permanent class was found to have the lowest prevalence of poor self-reported health in both samples: `r prettyNum(non_perm_srh_prev_a, big.mark=",")`% in Sample A and `r prettyNum(non_perm_srh_prev_b, big.mark=",")`% in Sample B (Figure x). In neither sample was the prevalence found to be different from that for the permanent employment group (`r prettyNum(perm_srh_prev_a, big.mark=",")`% and `r prettyNum(perm_srh_prev_b, big.mark=",")`% respectively). 

In Sample A, prevalence of common mental health disorder (measured by GHQ-12 caseness) was found to be higher among the non-permanent employment class than the permanent employment class: `r prettyNum(non_perm_ghq_prev_a)`% and `r prettyNum(perm_ghq_prev_a)`% respectively (Figure x). However, no different was observed between these classes in Sample B:  `r prettyNum(non_perm_ghq_prev_b)`% and `r prettyNum(perm_ghq_prev_b)`% respectively.

\newpage

**Figure x: Prevalence of poor self-reported health by employment contract class**
```{r, echo=FALSE}
image_read("../output/weighted/emp_contract_srh_dsr_grouped.tiff")

```

\newpage

**Figure x: Prevalence of common mental health disorder by employment contract class**
```{r, echo=FALSE}
image_read("../output/weighted/emp_contract_ghq_dsr_grouped.tiff")

```

\newpage

### Broken employment spells
```{r, echo=FALSE, include=FALSE}
### % poor SRH
## broken spells sample A
broken_srh_prev_a <- svy_dsr_df2$value[svy_dsr_df2$sample_grp=="A" &
                                       svy_dsr_df2$outcome_lab=="poor self-rated health" &
                                       svy_dsr_df2$sex_dv=="both" &
                                       svy_dsr_df2$class_mem=="Broken employment"]

## broken spells sample B
broken_srh_prev_b <- svy_dsr_df2$value[svy_dsr_df2$sample_grp=="B" &
                                       svy_dsr_df2$outcome_lab=="poor self-rated health" &
                                       svy_dsr_df2$sex_dv=="both" &
                                       svy_dsr_df2$class_mem=="Broken employment"]

## unbroken spells sample A
unbroken_srh_prev_a <- svy_dsr_df2$value[svy_dsr_df2$sample_grp=="A" &
                                       svy_dsr_df2$outcome_lab=="poor self-rated health" &
                                       svy_dsr_df2$sex_dv=="both" &
                                       svy_dsr_df2$class_mem=="Unbroken employment"]

## unbroken spells sample B
unbroken_srh_prev_b <- svy_dsr_df2$value[svy_dsr_df2$sample_grp=="B" &
                                       svy_dsr_df2$outcome_lab=="poor self-rated health" &
                                       svy_dsr_df2$sex_dv=="both" &
                                       svy_dsr_df2$class_mem=="Unbroken employment"]


### % poor mental health
## broken spells sample A
broken_ghq_prev_a <- svy_dsr_df2$value[svy_dsr_df2$sample_grp=="A" &
                                       svy_dsr_df2$outcome_lab=="poor mental health" &
                                       svy_dsr_df2$sex_dv=="both" &
                                       svy_dsr_df2$class_mem=="Broken employment"]

## broken spells sample B
broken_ghq_prev_b <- svy_dsr_df2$value[svy_dsr_df2$sample_grp=="B" &
                                       svy_dsr_df2$outcome_lab=="poor mental health" &
                                       svy_dsr_df2$sex_dv=="both" &
                                       svy_dsr_df2$class_mem=="Broken employment"]

## unbroken spells sample A
unbroken_ghq_prev_a <- svy_dsr_df2$value[svy_dsr_df2$sample_grp=="A" &
                                       svy_dsr_df2$outcome_lab=="poor mental health" &
                                       svy_dsr_df2$sex_dv=="both" &
                                       svy_dsr_df2$class_mem=="Unbroken employment"]

## unbroken spells sample B
unbroken_ghq_prev_b <- svy_dsr_df2$value[svy_dsr_df2$sample_grp=="B" &
                                       svy_dsr_df2$outcome_lab=="poor mental health" &
                                       svy_dsr_df2$sex_dv=="both" &
                                       svy_dsr_df2$class_mem=="Unbroken employment"]


```
The prevalence of poor self-rated health for the broken employment class in Sample A was `r prettyNum(broken_srh_prev_a)`% compared with `r prettyNum(broken_srh_prev_b)`% in sample B (Figure x). For both samples the prevalence among the broken employment class was slightly higher than the equivalent unbroken employment class - `r prettyNum(unbroken_srh_prev_a)`% and `r prettyNum(unbroken_srh_prev_b)`% respectively.

The prevalence of common mental health disorder among the broken employment class was `r prettyNum(broken_ghq_prev_a)`% in Sample A and `r prettyNum(broken_ghq_prev_b)`% in Sample B (Figure x). Similar prevalence rates were observed for the unbroken employment class - `r prettyNum(unbroken_ghq_prev_a)`% and `r prettyNum(unbroken_ghq_prev_b)`% respectively.

**Figure x: Prevalence of poor self-reported health by employment spells class**
```{r, echo=FALSE}
image_read("../output/weighted/emp_spells_srh_dsr_grouped.tiff")

```

\newpage

**Figure x: Prevalence of common mental health disorder by employment spells class**
```{r, echo=FALSE}
image_read("../output/weighted/emp_spells_ghq_dsr_grouped.tiff")

```

\newpage

### Multiple employment
```{r, echo=FALSE, include=FALSE}
### % poor SRH
## multiple emp sample A
multiple_srh_prev_a <- svy_dsr_df3$value[svy_dsr_df3$sample_grp=="A" &
                                       svy_dsr_df3$outcome_lab=="poor self-rated health" &
                                       svy_dsr_df3$sex_dv=="both" &
                                       svy_dsr_df3$class_mem=="Multiple employment"]

## multiple emp sample B
multiple_srh_prev_b <- svy_dsr_df3$value[svy_dsr_df3$sample_grp=="B" &
                                       svy_dsr_df3$outcome_lab=="poor self-rated health" &
                                       svy_dsr_df3$sex_dv=="both" &
                                       svy_dsr_df3$class_mem=="Multiple employment"]

## single emp sample A
single_srh_prev_a <- svy_dsr_df3$value[svy_dsr_df3$sample_grp=="A" &
                                       svy_dsr_df3$outcome_lab=="poor self-rated health" &
                                       svy_dsr_df3$sex_dv=="both" &
                                       svy_dsr_df3$class_mem=="Single employment"]

## single emp sample B
single_srh_prev_b <- svy_dsr_df3$value[svy_dsr_df3$sample_grp=="B" &
                                       svy_dsr_df3$outcome_lab=="poor self-rated health" &
                                       svy_dsr_df3$sex_dv=="both" &
                                       svy_dsr_df3$class_mem=="Single employment"]


### % poor mental health
## multiple emp sample A
multiple_ghq_prev_a <- svy_dsr_df3$value[svy_dsr_df3$sample_grp=="A" &
                                       svy_dsr_df3$outcome_lab=="poor mental health" &
                                       svy_dsr_df3$sex_dv=="both" &
                                       svy_dsr_df3$class_mem=="Multiple employment"]

## multiple emp sample B
multiple_ghq_prev_b <- svy_dsr_df3$value[svy_dsr_df3$sample_grp=="B" &
                                       svy_dsr_df3$outcome_lab=="poor mental health" &
                                       svy_dsr_df3$sex_dv=="both" &
                                       svy_dsr_df3$class_mem=="Multiple employment"]

## single emp sample A
single_ghq_prev_a <- svy_dsr_df3$value[svy_dsr_df3$sample_grp=="A" &
                                       svy_dsr_df3$outcome_lab=="poor mental health" &
                                       svy_dsr_df3$sex_dv=="both" &
                                       svy_dsr_df3$class_mem=="Single employment"]

## single emp sample B
single_ghq_prev_b <- svy_dsr_df3$value[svy_dsr_df3$sample_grp=="B" &
                                       svy_dsr_df3$outcome_lab=="poor mental health" &
                                       svy_dsr_df3$sex_dv=="both" &
                                       svy_dsr_df3$class_mem=="Single employment"]


```

The multiple employment class was found to have the lowest prevalence of poor self-reported health in both samples: `r prettyNum(multiple_srh_prev_a)`% in Sample A and `r prettyNum(multiple_srh_prev_b)`% in Sample B (Figure x). In both samples the prevalence of poor self-reported health was found to be lower than that for the single employment class (`r prettyNum(single_srh_prev_a)`% and `r prettyNum(single_srh_prev_b)`% respectively).

The Sample A multiple employment class was found to have a similar prevalence of common mental health disorder to the single employment class: `r prettyNum(multiple_ghq_prev_a)`% and `r prettyNum(single_ghq_prev_a)`% respectively (Figure x). However, the prevalence of common mental health disorder was higher among the Sample B multiple employment class compared to the single employment class: `r prettyNum(multiple_ghq_prev_b)`% and `r prettyNum(single_ghq_prev_b)`% respectively.

\newpage

**Figure x: Prevalence of poor self-reported health by multiple employment class**
```{r, echo=FALSE}
image_read("../output/weighted/multi_emp_srh_dsr_grouped.tiff")

```

\newpage

**Figure x: Prevalence of common mental health disorder by multiple employment class**

```{r, echo=FALSE}
image_read("../output/weighted/multi_emp_ghq_dsr_grouped.tiff")

```
